+++ changelog.md
@@ -1,3 +1,9 @@
+# 0.0.2
+
+* Added blacklist and config reload commands
+* Updated default blacklist
+* Namespaces can now be blacklisted
+
 # 0.0.1
 
 * Added tick scheduler